feat(num): Int is a first-class number everywhere — runtime/stdlib/contracts green
Make Value::Int accepted wherever a number was, and migrate the corpus:
- Central funnel: Value::as_f64() (widens Int) + as_int_exact() (indexing/counts);
  stdlib want_number delegates to it.
- Ranges/step (shared, byte-identical): int bounds+step → Int sequence, any float →
  float; fixes the '1..10 bounds must be numbers' regression. Indexing accepts Int;
  non-integral float index → Tier-2 'array index must be an int, got float'.
- Type::Number contract accepts both subtypes (params/returns/fields).
- Stdlib breadth: json/cbor/msgpack/otlp serialize Int; sql/redis bind+decode ints;
  net/http config readers, schema constraints, lru/array/string/bytes/set, decimal,
  convert, stream, tui — all accept Int. Fixed many silently-dropped-Int config bugs.
- Gate-7 migration: 3 goldens (monte-carlo exact LCG, data.as int division, schema
  'got int') + unit-test expectations (division-is-int, type()->int/float, truthiness).

VERIFIED: build + clippy --all-targets clean BOTH configs; vm_differential 354/0 (no
engine divergence); cargo test + --no-default-features fully green (lib 1974/0).

NUM runtime integration. Co-Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>

+++ src/stdlib/ai/response.rs
@@ -190,7 +190,8 @@ pub(crate) fn error_to_value(err: &genai::Error) -> Value {
     let (message, status) = describe_error(err);
     map.insert("message".to_string(), Value::Str(message.into()));
     if let Some(s) = status {
-        map.insert("status".to_string(), Value::Float(s as f64));
+        // NUM §4: an HTTP status code is an `Int`.
+        map.insert("status".to_string(), Value::Int(i64::from(s)));
     }
     Value::Object(crate::value::ObjectCell::new(map))
 }
